完善 控制台和完整的 insert, update, delete 操作。 1. 支持自动增长的值。 2. 支持逻辑删除。 3. 支持自动过滤掉删除的值。

+++ lightbatis-samples/src/main/java/titan/lightbatis/sample/mapper/MemberMapper.java
@@ -3,12 +3,13 @@ package titan.lightbatis.sample.mapper;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.Path;
 import com.querydsl.core.types.Predicate;
-import titan.lightbatis.mapper.LightbatisMapper;
-import titan.lightbatis.result.Page;
 import titan.lightbatis.result.PageList;
-import titan.lightbatis.sample.domain.Member;
+import titan.lightbatis.result.Page;
+import titan.lightbatis.mapper.LightbatisMapper;
+import titan.lightbatis.sample.model.entity.Member;
 
 import java.util.List;
+
 
 /**
  * MemberMapper 企业表 Mapper
@@ -29,7 +30,7 @@ public interface MemberMapper extends LightbatisMapper<Member> {
      * @param kindId
      * @return
      */
-    List<Member> listMemberByKindId(Short kindId);
+    List<Member> listMemberByKindId(Integer kindId);
 
     /**
      * 列出所有的企业
@@ -73,4 +74,3 @@ public interface MemberMapper extends LightbatisMapper<Member> {
      */
     public List<Member> listMemberFields(Path... paths);
 }
-
